requirements.txt
alembic==1.4.3
amqp==2.6.1
aniso8601==8.1.0
appdirs==1.4.3
attrs==20.3.0
black==20.8b1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
greenlet==0.4.17
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mock==4.0.3
msgpack==0.6.2
mypy-extensions==0.4.3
nameko==2.13.0
packaging==20.3
path==15.0.1
path.py==12.5.0
pathspec==0.8.1
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.10.0
pyparsing==2.4.6
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
